3D double to float

+++ korge/src/commonMain/kotlin/com/soywiz/korge3d/Light3D.kt
@@ -6,27 +6,27 @@ import com.soywiz.korma.geom.*
 @Korge3DExperimental
 fun Container3D.light(
 	color: RGBA = Colors.WHITE,
-	constantAttenuation: Double = 1.0,
-	linearAttenuation: Double = 0.0,
-	quadraticAttenuation: Double = 0.00111109,
+	constantAttenuation: Float = 1f,
+	linearAttenuation: Float = 0f,
+	quadraticAttenuation: Float = 0.00111109f,
 	callback: Light3D.() -> Unit = {}
 ) = Light3D(color, constantAttenuation, linearAttenuation, quadraticAttenuation).addTo(this, callback)
 
 @Korge3DExperimental
 open class Light3D(
 	var color: RGBA = Colors.WHITE,
-	var constantAttenuation: Double = 1.0,
-	var linearAttenuation: Double = 0.0,
-	var quadraticAttenuation: Double = 0.00111109
+	var constantAttenuation: Float = 1f,
+	var linearAttenuation: Float = 0f,
+	var quadraticAttenuation: Float = 0.00111109f
 ) : View3D() {
 	internal val colorVec = Vector3D()
 	internal val attenuationVec = Vector3D()
 
 	fun setTo(
 		color: RGBA = Colors.WHITE,
-		constantAttenuation: Double = 1.0,
-		linearAttenuation: Double = 0.0,
-		quadraticAttenuation: Double = 0.00111109
+		constantAttenuation: Float = 1f,
+		linearAttenuation: Float = 0f,
+		quadraticAttenuation: Float = 0.00111109f
 	): Light3D {
 		this.color = color
 		this.constantAttenuation = constantAttenuation
